ThemeLoader widget
This is an experiment in live-reloading themes from disk.

Useage is hopfully explained by the docs in theme_loader/mod.rs.
The basic idea is that you combine a set of keys into a 'theme'
with a macro, and then pass that theme and a path to a ThemeLoader
widget. This widget then loads the file at that path, and attempts
to parse out the expected keys, adding them to the env if successful.

For live-reloading, you need to enable the 'notify' feature.