Remove buggy on_monitor method and handle scrolling
on_monitor() does not work well with two monitors. If the parent window is moved by x, the tooltip gets moved by 2x.
Handling scrolling is necessary to avoid the tool tip gettings stuck in the showing state when the widget loses hot status due to scrolling and not mouse move.